+++ problem-spring-webflux/src/main/java/org/zalando/problem/spring/webflux/advice/ProblemHandling.java
@@ -4,8 +4,6 @@ import org.apiguardian.api.API;
 import org.zalando.problem.spring.webflux.advice.custom.CustomAdviceTrait;
 import org.zalando.problem.spring.webflux.advice.general.GeneralAdviceTrait;
 import org.zalando.problem.spring.webflux.advice.http.HttpAdviceTrait;
-import org.zalando.problem.spring.webflux.advice.io.IOAdviceTrait;
-import org.zalando.problem.spring.webflux.advice.routing.RoutingAdviceTrait;
 import org.zalando.problem.spring.webflux.advice.security.SecurityAdviceTrait;
 import org.zalando.problem.spring.webflux.advice.validation.ValidationAdviceTrait;
 
@@ -24,8 +22,6 @@ import static org.apiguardian.api.API.Status.STABLE;
  * @see CustomAdviceTrait
  * @see GeneralAdviceTrait
  * @see HttpAdviceTrait
- * @see IOAdviceTrait
- * @see RoutingAdviceTrait
  * @see SecurityAdviceTrait
  * @see ValidationAdviceTrait
  */
@@ -33,8 +29,6 @@ import static org.apiguardian.api.API.Status.STABLE;
 public interface ProblemHandling extends
         GeneralAdviceTrait,
         HttpAdviceTrait,
-        IOAdviceTrait,
-        RoutingAdviceTrait,
         SecurityAdviceTrait,
         ValidationAdviceTrait {
