--- problem-spring-webflux/src/main/java/org/zalando/problem/spring/webflux/advice/io/MessageNotReadableAdviceTrait.java
@@ -1,31 +1,0 @@
-package org.zalando.problem.spring.webflux.advice.io;
-
-import org.apiguardian.api.API;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.HttpMessageNotReadableException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.server.ServerWebExchange;
-import org.zalando.problem.Problem;
-import org.zalando.problem.Status;
-import org.zalando.problem.spring.webflux.advice.AdviceTrait;
-import reactor.core.publisher.Mono;
-
-import static org.apiguardian.api.API.Status.INTERNAL;
-import static org.apiguardian.api.API.Status.STABLE;
-
-/**
- * @see HttpMessageNotReadableException
- * @see Status#BAD_REQUEST
- */
-@API(status = STABLE)
-public interface MessageNotReadableAdviceTrait extends AdviceTrait {
-
-    @API(status = INTERNAL)
-    @ExceptionHandler
-    default Mono<ResponseEntity<Problem>> handleMessageNotReadableException(
-            final HttpMessageNotReadableException exception,
-            final ServerWebExchange request) {
-        return create(Status.BAD_REQUEST, exception, request);
-    }
-
-}
