--- problem-spring-webflux/src/main/java/org/zalando/problem/spring/webflux/advice/io/MultipartAdviceTrait.java
@@ -1,27 +1,0 @@
-package org.zalando.problem.spring.webflux.advice.io;
-
-import org.apiguardian.api.API;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.multipart.MultipartException;
-import org.springframework.web.server.ServerWebExchange;
-import org.zalando.problem.Problem;
-import org.zalando.problem.Status;
-import org.zalando.problem.spring.webflux.advice.AdviceTrait;
-import reactor.core.publisher.Mono;
-
-import static org.apiguardian.api.API.Status.INTERNAL;
-
-// TODO find a better name
-//@API(status = STABLE)
-public interface MultipartAdviceTrait extends AdviceTrait {
-
-    @API(status = INTERNAL)
-    @ExceptionHandler
-    default Mono<ResponseEntity<Problem>> handleMultipart(
-            final MultipartException exception,
-            final ServerWebExchange request) {
-        return create(Status.BAD_REQUEST, exception, request);
-    }
-
-}
