--- problem-spring-webflux/src/main/java/org/zalando/problem/spring/webflux/advice/io/package-info.java
@@ -1,5 +1,0 @@
-@ParametersAreNonnullByDefault
-package org.zalando.problem.spring.webflux.advice.io;
-
-import javax.annotation.ParametersAreNonnullByDefault;
-
