--- problem-spring-webflux/src/main/java/org/zalando/problem/spring/webflux/advice/routing/package-info.java
@@ -1,5 +1,0 @@
-@ParametersAreNonnullByDefault
-package org.zalando.problem.spring.webflux.advice.routing;
-
-import javax.annotation.ParametersAreNonnullByDefault;
-
